Refactor API Path
