+++ backend/src/main/java/com/bulletjournal/templates/controller/model/CategoryUnsubscribeParams.java
@@ -1,0 +1,23 @@
+package com.bulletjournal.templates.controller.model;
+
+import javax.validation.constraints.NotBlank;
+
+public class CategoryUnsubscribeParams {
+  @NotBlank
+  private Long selectionId;
+
+  public CategoryUnsubscribeParams() {
+  }
+
+  public CategoryUnsubscribeParams(Long selectionId) {
+    this.selectionId = selectionId;
+  }
+
+  public Long getSelectionId() {
+    return selectionId;
+  }
+
+  public void setSelectionId(Long selectionId) {
+    this.selectionId = selectionId;
+  }
+}
